requirements.txt
Flask==1.0.2
Flask-Script==2.0.6
python-docx==0.8.7
